Implemented the Java implementations for primitive and collection model instance elements.

+++ ocl20forEclipse/eclipse/tudresden.ocl20.pivot.modelbus/src/tudresden/ocl20/pivot/modelbus/modelinstance/types/IModelInstanceInteger.java
@@ -38,5 +38,5 @@ public interface IModelInstanceInteger extends IModelInstancePrimitiveType {
 	 * 
 	 * @return The {@link Long} value of this {@link IModelInstanceInteger}.
 	 */
-	Long getInteger();
+	Long getLong();
 }
